feat Tags: Create a Loader Component and create css element for block text.
