+++ izanami-frontend/src/App.css
@@ -48,6 +48,17 @@ header .nav-link {
   padding: 20px;
   margin-bottom: 12px;
 }
+.item-block {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 10%;
+}
+.item-text {
+  font-size: 21px;
+  margin-Bottom: 24px;
+}
 
 /*.fieldset-form fieldset:nth-of-type(2n) {
   background-color: #414140;
