+++ izanami-frontend/src/components/Loader.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { CSSProperties } from "react";
+import ClipLoader from "react-spinners/ClipLoader";
+
+const override: CSSProperties = {
+  display: "block",
+  margin: "0 auto",
+};
+
+export function Loader() {
+  return (
+    <ClipLoader
+      color="#fff"
+      cssOverride={override}
+      size={50}
+      aria-label="Loading Spinner"
+      data-testid="loader"
+    />
+  );
+}
